Refactor the directories
