--- website/src/playground/types/index.ts
@@ -1,3 +1,0 @@
-import { editor } from 'monaco-editor';
-
-export type Diagnostic = Readonly<editor.IMarkerData>;
